% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task2_MarianneSeijo.R
\name{popmean}
\alias{popmean}
\title{popmean.}
\usage{
popmean(dat, intervention, years)
}
\arguments{
\item{dat}{dataset with columns "zone",    "population"    "year"    "none"    "IRS"    "ITN"    "IRS+ITN"}

\item{intervention}{intervention concerned ("none"    "IRS"    "ITN"    "IRS+ITN")}

\item{years}{years concerned}
}
\value{
Returns the value of \code{arg1}
}
\description{
\code{popmean} calculates the national prevalence (population-weighted mean of the two zones) by years for an intervention type.
}
\examples{
{
dat = read.delim(text = "zone	population	year	none	IRS	ITN	IRS+ITN
Zone 1	692089	2005	0.41	0.40	0.39	0.40
Zone 1	692089	2006	0.37	0.35	0.35	0.35
Zone 1	692089	2007	0.31	0.29	0.29	0.29
Zone 2	324469	2005	0.26	0.26	0.26	0.26
Zone 2	324469	2006	0.21	0.20	0.20	0.20
Zone 2	324469	2007	0.16	0.15	0.15	0.15", sep = "\t", header = TRUE)
popmean( dat, intervention = "IRS", years = 2005:2006)
}

}
